requirements.txt
# Django и REST Framework
Django==4.2.29
djangorestframework==3.16.1

# JWT аутентификация
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0

# Документация (Swagger)
drf-spectacular==0.28.0

# Работа с картинками товаров
Pillow==10.4.0

# CORS - для работы с фронтендом
django-cors-headers==4.3.1

# Фильтрация товаров
django-filter==24.2

# Утилиты
asgiref==3.11.1
sqlparse==0.5.5
typing_extensions==4.15.0

# Для продакшна (потом пригодится)
gunicorn==22.0.0
